examples/mnist/requirements.txt
bitorch
bitorch_engine
